LayoutTests/svg/gradients/linear-gradient-default-length.html

<!DOCTYPE html> 
<title>SVGLinearGradientElement coordinate defaults</title>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<svg>
    <linearGradient id="lg"/>
</svg>
<script>
function testLengthValue(result, expected, testcase) {
    test(function() {
        assert_equals(result, expected);
    }, "Testing " + document.title + " " + testcase);
}
var linearGradient = document.querySelector("#lg");
testLengthValue(linearGradient.x1.baseVal.valueAsString, "0%", "x1");
testLengthValue(linearGradient.x2.baseVal.valueAsString, "100%", "x2");
testLengthValue(linearGradient.y1.baseVal.valueAsString, "0%", "y1");
testLengthValue(linearGradient.y2.baseVal.valueAsString, "0%", "y2");
</script>
